move countDocuments to collection_ops
