+++ lib/collection.js
@@ -22,6 +22,7 @@ const getReadPreference = require('./utils').getReadPreference;
 const bulkWrite = require('./operations/collection_ops').bulkWrite;
 const checkForAtomicOperators = require('./operations/collection_ops').checkForAtomicOperators;
 const count = require('./operations/collection_ops').count;
+const countDocuments = require('./operations/collection_ops').countDocuments;
 const createIndex = require('./operations/collection_ops').createIndex;
 const createIndexes = require('./operations/collection_ops').createIndexes;
 const decorateWithReadConcern = require('./operations/collection_ops').decorateWithReadConcern;
@@ -95,7 +96,7 @@ const mergeKeys = ['readPreference', 'ignoreUndefined'];
  * @property {object} hint Get current index hint for collection.
  * @return {Collection} a Collection instance.
  */
-const Collection = function(db, topology, dbName, name, pkFactory, options) {
+function Collection(db, topology, dbName, name, pkFactory, options) {
   checkCollectionName(name);
 
   // Unpack variables
@@ -174,7 +175,7 @@ const Collection = function(db, topology, dbName, name, pkFactory, options) {
     // Read Concern
     readConcern: options.readConcern
   };
-};
+}
 
 Object.defineProperty(Collection.prototype, 'dbName', {
   enumerable: true,
@@ -498,7 +499,7 @@ Collection.prototype.insertMany = function(docs, options, callback) {
       MongoError.create({ message: 'docs parameter must be an array of documents', driver: true })
     );
   } else if (!Array.isArray(docs)) {
-    return new this.s.promiseLibrary(function(resolve, reject) {
+    return new this.s.promiseLibrary((resolve, reject) => {
       reject(
         MongoError.create({ message: 'docs parameter must be an array of documents', driver: true })
       );
@@ -1384,43 +1385,12 @@ Collection.prototype.estimatedDocumentCount = function(options, callback) {
  */
 
 Collection.prototype.countDocuments = function(query, options, callback) {
-  var args = Array.prototype.slice.call(arguments, 0);
+  const args = Array.prototype.slice.call(arguments, 0);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
   query = args.length ? args.shift() || {} : {};
   options = args.length ? args.shift() || {} : {};
 
   return executeOperation(this.s.topology, countDocuments, [this, query, options, callback]);
-};
-
-const countDocuments = function(collection, query, options, callback) {
-  const skip = options.skip;
-  const limit = options.limit;
-  options = Object.assign({}, options);
-
-  const pipeline = [{ $match: query }];
-
-  // Add skip and limit if defined
-  if (typeof skip === 'number') {
-    pipeline.push({ $skip: skip });
-  }
-
-  if (typeof limit === 'number') {
-    pipeline.push({ $limit: limit });
-  }
-
-  pipeline.push({ $group: { _id: null, n: { $sum: 1 } } });
-
-  delete options.limit;
-  delete options.skip;
-
-  // TODO: look out for how this plays into operations refactor
-  collection.aggregate(pipeline, options, function(err, result) {
-    if (err) return handleCallback(callback, err);
-    result
-      .toArray()
-      .then(docs => handleCallback(callback, null, docs[0].n))
-      .catch(e => handleCallback(e));
-  });
 };
 
 /**
@@ -1687,8 +1657,6 @@ Collection.prototype.findAndRemove = function(query, sort, options, callback) {
  * @return {(null|AggregationCursor)}
  */
 Collection.prototype.aggregate = function(pipeline, options, callback) {
-  const self = this;
-
   if (Array.isArray(pipeline)) {
     // Set up callback if one is provided
     if (typeof options === 'function') {
@@ -1743,13 +1711,13 @@ Collection.prototype.aggregate = function(pipeline, options, callback) {
   if (
     pipeline.length > 0 &&
     pipeline[pipeline.length - 1]['$out'] &&
-    self.s.topology.capabilities().commandsTakeWriteConcern
+    this.s.topology.capabilities().commandsTakeWriteConcern
   ) {
-    applyWriteConcern(command, { db: self.s.db, collection: self }, options);
+    applyWriteConcern(command, { db: this.s.db, collection: this }, options);
   }
 
   // Have we specified collation
-  decorateWithCollation(command, self, options);
+  decorateWithCollation(command, this, options);
 
   // If we have bypassDocumentValidation set
   if (typeof options.bypassDocumentValidation === 'boolean') {
@@ -1758,7 +1726,7 @@ Collection.prototype.aggregate = function(pipeline, options, callback) {
 
   // Do we have a readConcern specified
   if (!ignoreReadConcern) {
-    decorateWithReadConcern(command, self, options);
+    decorateWithReadConcern(command, this, options);
   }
 
   // If we have allowDiskUse defined
