+++ lib/operations/collection_ops.js
@@ -210,6 +210,36 @@ function count(coll, query, options, callback) {
   });
 }
 
+function countDocuments(coll, query, options, callback) {
+  const skip = options.skip;
+  const limit = options.limit;
+  options = Object.assign({}, options);
+
+  const pipeline = [{ $match: query }];
+
+  // Add skip and limit if defined
+  if (typeof skip === 'number') {
+    pipeline.push({ $skip: skip });
+  }
+
+  if (typeof limit === 'number') {
+    pipeline.push({ $limit: limit });
+  }
+
+  pipeline.push({ $group: { _id: null, n: { $sum: 1 } } });
+
+  delete options.limit;
+  delete options.skip;
+
+  coll.aggregate(pipeline, options, (err, result) => {
+    if (err) return handleCallback(callback, err);
+    result
+      .toArray()
+      .then(docs => handleCallback(callback, null, docs[0].n))
+      .catch(e => handleCallback(e));
+  });
+}
+
 /**
  * Create an index on the db and collection.
  *
@@ -1394,6 +1424,7 @@ module.exports = {
   bulkWrite,
   checkForAtomicOperators,
   count,
+  countDocuments,
   createIndex,
   createIndexes,
   decorateWithCollation,
